+++ LabSources/SampleScripts/Azure/AzureStackHub.ps1
@@ -96,4 +96,4 @@ Add-LabMachineDefinition -Name ALWBOnAzS -Memory 4GB -DomainName contoso.com -Op
 
 Install-Lab
 
-Show-LabDeploymentSummary -Detailed
+Show-LabDeploymentSummary