+++ LabSources/SampleScripts/HyperV/Single 2012R2 Server DC SQL2014 Web CA.ps1
@@ -30,4 +30,4 @@ Install-LabSoftwarePackage -ComputerName (Get-LabVM) -Path $labSources\SoftwareP
 
 Enable-LabCertificateAutoenrollment -Computer -User -CodeSigning
 
-Show-LabDeploymentSummary -Detailed
+Show-LabDeploymentSummary