+++ LabSources/SampleScripts/Introduction/06 SQL Server and client, domain joined.ps1
@@ -21,4 +21,4 @@ Add-LabMachineDefinition -Name Client1 -OperatingSystem 'Windows 10 Pro'
 
 Install-Lab
 
-Show-LabDeploymentSummary -Detailed
+Show-LabDeploymentSummary