+++ LabSources/SampleScripts/Introduction/07.2 Exchange 2016 Server and client, domain joined.ps1
@@ -31,4 +31,4 @@ Restart-LabVM -ComputerName Lab2016EX1 -Wait
 
 Install-Lab -PostInstallations
 
-Show-LabDeploymentSummary -Detailed
+Show-LabDeploymentSummary