+++ LabSources/SampleScripts/Scenarios/CM-1902.ps1
@@ -495,5 +495,5 @@ if ($PostInstallations.IsPresent) {
 else {
     Install-Lab
 }
-Show-LabDeploymentSummary -Detailed
+Show-LabDeploymentSummary
 #endregion