+++ LabSources/SampleScripts/Scenarios/TFS 2015 Deployment.ps1
@@ -29,4 +29,4 @@ Add-LabMachineDefinition -Name tfsBuild1 -Roles $role
 
 Install-Lab
 
-Show-LabDeploymentSummary -Detailed
+Show-LabDeploymentSummary