+++ LabSources/SampleScripts/Workshops/ADRecoveryLab - HyperV.ps1
@@ -102,4 +102,4 @@ if ($createCheckpoint)
     Checkpoint-LabVM -All -SnapshotName 'AfterSetupComplete'
 }
 
-Show-LabDeploymentSummary -Detailed
+Show-LabDeploymentSummary